Added in event listeners to Weather.js and tested API functionality associatd
Could be a server issue but the date and time in FormatDate was not appearing or would appear then go back to a day array digit. Need to keep this in mind as I move through building the rest of the application. Tried removing JSX from FormatDate but this created similar issue at different caliber.